print offers bought+sort done
todo, SSE printen

+++ DataScienceKmeans/src/Cluster.java
@@ -7,16 +7,16 @@ import java.util.Map;
  * Created by Orlando on 2-3-2017.
  */
 public class Cluster {
-    private float[] center = new float[32]; //needs to be float for recalculating center
+    private double[] center = new double[32]; //needs to be float for recalculating center
     private int id;
     private List<Point> points = new ArrayList<>();
     private Map<Integer, Integer>offerBoughtCount= new HashMap<>();
 
-    public float[] getCenter() {
+    public double[] getCenter() {
         return center;
     }
 
-    public void setCenter(float[] center) {
+    public void setCenter(double[] center) {
         this.center = center;
     }
